Add tests based on the pool state
