+++ radix-engine-tests/src/lib.rs
@@ -1,2 +1,3 @@
 // Provides assets paths to benchmarks.
 pub mod common;
+pub mod pool_stubs;
